Move css and theme to new exported file
