+++ src/components/Button/index.ts
@@ -2,5 +2,6 @@ import Button from './Button';
 import ButtonMode from './ButtonMode';
 import ButtonText from './ButtonText';
 import ButtonOutline from './ButtonOutline';
+import { getTheme, ButtonTheme } from './ButtonTheme';
 
-export {Button, ButtonMode, ButtonText, ButtonOutline};
+export {Button, ButtonMode, ButtonText, ButtonOutline, ButtonTheme, getTheme};
